aider: Modified SQL query in '/writers' route to select records based on 'UserID' instead of 'Username'.
# Context:
USER: OperationalError
MySQLdb.OperationalError: (1054, "Unknown column 'Username' in 'where clause'") whenloading the writers.html page
ASSISTANT: The error message indicates that the column 'Username' does not exist in the 'writers' table. Since we have changed the 'writers' table to use 'UserID' instead of 'Username', we need to modify the SQL query in the '/writers' route in the `app.py` file to select records based on 'UserID' instead of 'Username'.

Here is the change:

```python
app.py
<<<<<<< ORIGINAL
@app.route('/writers')
def writers():
    if 'username' not in session:
        return redirect(url_for('login'))

    from MySQLdb.cursors import DictCursor
    cursor = mysql.connection.cursor(DictCursor)
    cursor.execute(''' SELECT * FROM writers WHERE Username = %s ''', (session['username'],))
    writers = cursor.fetchall()
    cursor.close()

    return render_template('writers.html', writers=writers)
=======
@app.route('/writers')
def writers():
    if 'username' not in session:
        return redirect(url_for('login'))

    from MySQLdb.cursors import DictCursor
    cursor = mysql.connection.cursor(DictCursor)
    cursor.execute(''' SELECT * FROM users WHERE username = %s ''', (session['username'],))
    user = cursor.fetchone()
    cursor.execute(''' SELECT * FROM writers WHERE UserID = %s ''', (user['UserID'],))
    writers = cursor.fetchall()
    cursor.close()

    return render_template('writers.html', writers=writers)
>>>>>>> UPDATED
```

In the updated code, we first select the user record based on the username stored in the session. Then we use the 'UserID' of the selected user to select the associated records from the 'writers' table.