aider: Removed fields that are not in the `projects` table from the form in `add_project.html`.
# Context:
USER: update add_projects.html to only support the fields in the projects table
ASSISTANT: To update the `add_project.html` file, we need to remove the fields that are not in the `projects` table.

However, I don't have the `add_project.html` file in the list of *read-write* files you provided. Could you please make it *read-write* and provide its current content?
USER: I added these *read-write* files: templates/add_project.html
ASSISTANT: Great, now that we have the `add_project.html` file in *read-write* mode, we can proceed with the changes.

From the provided `add_project.html` file, it seems that the fields `task`, `topic`, `style`, `audience`, `length`, `format`, and `additional_information` are not in the `projects` table. We need to remove these fields from the form.

Here is the *edit block* for the changes:

```html
templates/add_project.html
<<<<<<< ORIGINAL
        <label for="task">Task:</label><br>
        <input type="text" id="task" name="task"><br>
        <label for="topic">Topic:</label><br>
        <input type="text" id="topic" name="topic"><br>
        <label for="style">Style:</label><br>
        <input type="text" id="style" name="style"><br>
        <label for="audience">Audience:</label><br>
        <input type="text" id="audience" name="audience"><br>
        <label for="length">Length:</label><br>
        <input type="text" id="length" name="length"><br>
        <label for="format">Format:</label><br>
        <input type="text" id="format" name="format"><br>
        <label for="additional_information">Additional Information:</label><br>
        <textarea id="additional_information" name="additional_information"></textarea><br>
=======
>>>>>>> UPDATED
```

This *edit block* removes the fields that are not in the `projects` table from the form in `add_project.html`.