aider: Add a new button that, when clicked, will redirect the user to the writers page.
# Context:
USER: make a new button similar to 'Open Projects' call 'Open Writers' which will route to the writers.html file
ASSISTANT: To add a new button that routes to the writers.html file, you need to modify the home.html file in the templates directory.

Here is the change you need to make:

```html
templates\home.html
<<<<<<< ORIGINAL
    {% include 'header.html' %}
    <h1>Welcome to the Home Page!</h1>
    <button onclick="location.href='projects'">Open Projects</button>
=======
    {% include 'header.html' %}
    <h1>Welcome to the Home Page!</h1>
    <button onclick="location.href='projects'">Open Projects</button>
    <button onclick="location.href='writers'">Open Writers</button>
>>>>>>> UPDATED
```

This change adds a new button that, when clicked, will redirect the user to the writers page.