aider: Replaced the 'Add Writer' button with a form that includes fields for each column in the writers table on add_writer.html.
# Context:
USER: I added these *read-write* files: templates/header.html, templates/add_writer.html, templates/writers.html
ASSISTANT: Great! Now that you have added the necessary files, you can proceed with making changes to them. If you need any assistance with modifying the code in these files, feel free to ask.
USER: remove the 'add writer' button on add_writer.html and replace witha form that will allow the creation of a new record in the writers table in the database
ASSISTANT: To replace the 'Add Writer' button with a form on the add_writer.html page, you need to modify the add_writer.html file in the templates directory.

Here is the change you need to make:

```html
templates\add_writer.html
<<<<<<< ORIGINAL
<!DOCTYPE html>
<html>
<head>
    <title>Add Writer</title>
</head>
<body>
    {% include 'header.html' %}
    <h1>Add Writer</h1>
    <form action="/add_writer" method="post">
        <!-- Add form fields for each field in the writers table -->
        <input type="submit" value="Add Writer">
    </form>
</body>
</html>
=======
<!DOCTYPE html>
<html>
<head>
    <title>Add Writer</title>
</head>
<body>
    {% include 'header.html' %}
    <h1>Add Writer</h1>
    <form action="/add_writer" method="post">
        <label for="username">Username:</label><br>
        <input type="text" id="username" name="username"><br>
        <label for="task">Task:</label><br>
        <input type="text" id="task" name="task"><br>
        <label for="topic">Topic:</label><br>
        <input type="text" id="topic" name="topic"><br>
        <label for="style">Style:</label><br>
        <input type="text" id="style" name="style"><br>
        <label for="audience">Audience:</label><br>
        <input type="text" id="audience" name="audience"><br>
        <label for="length">Length:</label><br>
        <input type="text" id="length" name="length"><br>
        <label for="format">Format:</label><br>
        <input type="text" id="format" name="format"><br>
        <label for="additional_information">Additional Information:</label><br>
        <input type="text" id="additional_information" name="additional_information"><br>
        <input type="submit" value="Add Writer">
    </form>
</body>
</html>
>>>>>>> UPDATED
```

This change replaces the 'Add Writer' button with a form that includes fields for each column in the writers table. When the form is submitted, the data will be sent to the server using a POST request. The server will then need to handle this request and insert the new writer into the database.