aider: Added form fields for each writer attribute in the `edit_writer.html` file.
# Context:
USER: I added these *read-write* files: templates/edit_writer.html
ASSISTANT: Great! Now that we have the `templates/edit_writer.html` file set to read-write, we can proceed with the necessary changes.

Assuming that you want to make all the fields of a writer editable, we need to add form fields for each writer attribute in the `edit_writer.html` file.

Here is the *edit block* for this change:

```html
templates\edit_writer.html
<<<<<<< ORIGINAL
    <form action="/edit_writer/{{ writer.id }}" method="post">
        <!-- Add your form fields here -->
        <input type="submit" value="Update Writer">
    </form>
=======
    <form action="/edit_writer/{{ writer.id }}" method="post">
        <label for="task">Task:</label><br>
        <input type="text" id="task" name="task" value="{{ writer.Task }}"><br>
        <label for="topic">Topic:</label><br>
        <input type="text" id="topic" name="topic" value="{{ writer.Topic }}"><br>
        <label for="style">Style:</label><br>
        <input type="text" id="style" name="style" value="{{ writer.Style }}"><br>
        <label for="audience">Audience:</label><br>
        <input type="text" id="audience" name="audience" value="{{ writer.Audience }}"><br>
        <label for="length">Length:</label><br>
        <input type="text" id="length" name="length" value="{{ writer.Length }}"><br>
        <label for="format">Format:</label><br>
        <input type="text" id="format" name="format" value="{{ writer.Format }}"><br>
        <label for="additional_information">Additional Information:</label><br>
        <input type="text" id="additional_information" name="additional_information" value="{{ writer.AdditionalInformation }}"><br>
        <input type="submit" value="Update Writer">
    </form>
>>>>>>> UPDATED
```

This change adds form fields for each writer attribute. The value of each form field is set to the current value of the corresponding writer attribute. When the form is submitted, these values will be sent in the POST request to the server.

Next, we need to update the `edit_writer` route in `app.py` to handle the POST request and update the writer's data in the database. Could you please add the `app.py` file to the *read-write* files?