aider: Changed the edit link in the `writers.html` template to use the `WriterID` field instead of `id`.
# Context:
USER: check the fields but I think the variable is wrong:
ASSISTANT: I see, you suspect that the variable used in the URL for the edit link might be incorrect. Let's check the fields in the `writers` table to confirm the correct variable name.

In the `app.py` file, the `writers` route fetches data from the `writers` table in the database. The fetched data is stored in the `writers` variable, which is then passed to the `writers.html` template.

In the `writers.html` template, we loop over the `writers` variable to display each writer's information. The edit link for each writer is created using the `writer.id` variable.

If the `id` field does not exist in the `writers` table, then the `writer.id` variable will be undefined, and the edit link will not work correctly.

To confirm the correct variable name, we need to check the fields in the `writers` table. If the `id` field does not exist, we need to find the correct field that uniquely identifies each writer and use that field in the URL for the edit link.

Please provide the fields in the `writers` table so we can confirm the correct variable name.
USER: I added these *read-write* files: templates/writers.html
ASSISTANT: Thank you for providing the `writers.html` file. Based on the provided file, it seems that the `id` field is being used to create the edit link. However, if the `id` field is not present in the `writers` table, this would cause an issue.

To confirm, could you please provide the structure of the `writers` table? Specifically, we need to know the names of the fields in the `writers` table. This will help us determine if the `id` field exists, or if we should be using a different field to create the edit link.
USER: I added these *read-write* files: templates/writers.html
ASSISTANT: Thank you for providing the `writers.html` file. Based on the provided file, it seems that the `id` field is being used to create the edit link. However, if the `id` field is not present in the `writers` table, this would cause an issue.

To confirm, could you please provide the structure of the `writers` table? Specifically, we need to know the names of the fields in the `writers` table. This will help us determine if the `id` field exists, or if we should be using a different field to create the edit link.
USER: | WriterID | UserID | Task            | Topic                                                 | Style                                  | Audience                                 | Length     | Format   | AdditionalInformation
ASSISTANT: Thank you for providing the structure of the `writers` table. It appears that the unique identifier for each writer is `WriterID`, not `id`.

We need to modify the `templates\writers.html` file to use `WriterID` instead of `id` in the URL for the edit link.

Here is the change:

```html
templates\writers.html
<<<<<<< ORIGINAL
<td style="padding: 15px; text-align: left;"><a href="/edit_writer/{{ writer.id }}">Edit</a></td>
=======
<td style="padding: 15px; text-align: left;"><a href="/edit_writer/{{ writer.WriterID }}">Edit</a></td>
>>>>>>> UPDATED
```

This change will ensure that the correct writer id is passed to the edit_writer route when the Edit link is clicked.